Tutorial Complete
Not quite 100% functional, but as far as I plan to take it. Feel free to play around and take a look at what I've done. All of it should be under my folder and you can play it using the play button on the top right (currently the game is the default scene). We can get rid of it whenever, but it was a good learning experience.